edit

+++ main.h
@@ -13,6 +13,6 @@ int printf_integer(va_list args, int len);
 int _printf(const char *format, ...);
 int getNumberLength(int myNum);
 void check_number(int n);
-int dtob(unsigned int n, int len);
+int dtob(unsigned long int n, int len);
 
 #endif
